Added routes and controllers for delete user, that deletes every comment, likes, photos, posts from the database

+++ controllers/posts.js
@@ -1,95 +1,96 @@
 import Post from "../models/Posts.js";
 import User from "../models/User.js";
 import Comment from "../models/Comment.js";
+import fs from "fs"
 
 // Create
-export const createPost = async (req,res)=>{
+export const createPost = async (req, res) => {
     try {
-        const{userId,description,picturePath,location}=req.body;
-        const user=await User.findById(userId);
+        const { userId, description, picturePath, location } = req.body;
+        const user = await User.findById(userId);
         const newPost = new Post({
             userId,
             firstName: user.firstName,
             lastName: user.lastName,
-            location:location!="" ? location : user.location,
+            location: location != "" ? location : user.location,
             description,
-            userPicturePath:user.picturePath,
+            userPicturePath: user.picturePath,
             picturePath,
-            likes:{}
+            likes: {}
         })
         await newPost.save();
 
-        const post=await Post.find();
+        const post = await Post.find();
         res.status(201).json(post);
 
     } catch (error) {
-        res.status(409).json({message:error.message})
+        res.status(409).json({ message: error.message })
     }
 };
 
 // Read
-export const getFeedPosts=async (req,res)=>{
+export const getFeedPosts = async (req, res) => {
     try {
-        const post=await Post.find();
+        const post = await Post.find();
         res.status(200).json(post);
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
 
-export const getUserPosts= async (req,res)=>{
+export const getUserPosts = async (req, res) => {
     try {
-        const {userId}=req.params;
-        const post =await Post.find({userId});
+        const { userId } = req.params;
+        const post = await Post.find({ userId });
         res.status(200).json(post);
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
-export const getPost= async (req,res)=>{
+export const getPost = async (req, res) => {
     try {
-        const {id}=req.params;
-        const post =await Post.findById(id).populate('comments')
+        const { id } = req.params;
+        const post = await Post.findById(id).populate('comments')
         res.status(200).json(post);
-        
+
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
 
 // Update
-export const likePost= async (req,res)=>{
+export const likePost = async (req, res) => {
     try {
-        const {id}=req.params;
-        const {userId}=req.body;
-        const post =await Post.findById(id);
+        const { id } = req.params;
+        const { userId } = req.body;
+        const post = await Post.findById(id);
         const isLiked = post.likes.get(userId);
 
-        if(isLiked){
+        if (isLiked) {
             post.likes.delete(userId);
         }
-        else{
+        else {
             post.likes.set(userId, true)
         }
 
-        const updatedPost= await Post.findByIdAndUpdate(
+        const updatedPost = await Post.findByIdAndUpdate(
             id,
-            {likes: post.likes},
-            {new : true}
+            { likes: post.likes },
+            { new: true }
         )
         res.status(200).json(updatedPost);
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
 
-export const commentPost=async (req,res)=>{
+export const commentPost = async (req, res) => {
     try {
-        const {id}=req.params;
-        const {userId,firstName,lastName,userPicturePath}=req.body;
-        const post =await Post.findById(id).populate('comments');
+        const { id } = req.params;
+        const { userId, firstName, lastName, userPicturePath } = req.body;
+        const post = await Post.findById(id).populate('comments');
 
-        const comment=new Comment({
+        const comment = new Comment({
             userId: userId,
             firstName: firstName,
             lastName: lastName,
@@ -102,34 +103,49 @@ export const commentPost=async (req,res)=>{
         res.status(200).json(post);
 
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
 
 
-export const commentDelete=async (req,res)=>{
+export const commentDelete = async (req, res) => {
     try {
-        const {id,commentId}=req.params;
+        const { id, commentId } = req.params;
         await Comment.findByIdAndDelete(commentId)
-        const post =await Post.findById(id).populate('comments');
+        const post = await Post.findById(id).populate('comments');
         await post.comments.remove(commentId)
         await post.save()
         res.status(200).json(post);
 
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
 
 
-export const deletePost=async (req,res)=>{
+export const deletePost = async (req, res) => {
     try {
-        const {id}=req.params;
-        await Post.findByIdAndDelete(id)
-        const posts=await Post.find()
-        res.status(200).json(posts);
+        const { id } = req.params;
+        const { picturePath } = req.body;
+        const directoryPath = "public/assets/";
+        const exists = fs.existsSync(`${directoryPath}${picturePath}`)
+
+
+        if (exists) {
+            fs.unlink(`${directoryPath}${picturePath}`, async (err) => {
+                if (err) {
+                    res.status(500).send({
+                        message: "Could not delete the file. " + err,
+                    });
+                }
+
+                await Post.findByIdAndDelete(id)
+                const posts = await Post.find()
+                res.status(200).json(posts);
+            })
+        }
 
     } catch (error) {
-        res.status(404).json({message:error.message})
+        res.status(404).json({ message: error.message })
     }
 }
